# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TasteTracker API",
        "description": "API REST del restaurante para gestionar menú, pedidos, clientes y empleados.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/v1/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar todos los pedidos",
                "description": "Obtiene todos los pedidos con información del cliente y empleado (sin detalles).",
                "operationId": "763c73f4fe562b79e809f8a4e332c4a8",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pedido"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un nuevo pedido con sus detalles",
                "description": "Crea un pedido y sus detalles en una transacción atómica. Si algún detalle falla, todo se revierte.",
                "operationId": "9ac38a0301bf52d679a30562d7588577",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/pedidos/{pedido}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener un pedido con sus relaciones",
                "description": "Devuelve un pedido con cliente, empleado y detalles (incluyendo el menú de cada detalle).",
                "operationId": "f1ab44575bbc487b45c15c364af56db1",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar el estado de un pedido",
                "description": "Actualiza únicamente el estado del pedido (id_estado).",
                "operationId": "98b0201bcc0ca983beec1ffed0716abe",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_estado"
                                ],
                                "properties": {
                                    "id_estado": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Eliminar un pedido",
                "description": "Elimina un pedido. Los detalles asociados se eliminan en cascada.",
                "operationId": "ba0fc0d0032163db9e5decf2deaad517",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar parcialmente un pedido (estado)",
                "description": "Actualiza el estado del pedido (id_estado).",
                "operationId": "96093dafbe457c174a4681f33d9c2cdc",
                "parameters": [
                    {
                        "name": "pedido",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_estado"
                                ],
                                "properties": {
                                    "id_estado": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "required": [
                    "nombre_cliente"
                ],
                "properties": {
                    "id_cliente": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre_cliente": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 1 #123, Col. Centro"
                    }
                },
                "type": "object"
            },
...
            "Empleado": {
                "required": [
                    "nombre_completo",
                    "rol"
                ],
                "properties": {
                    "id_empleado": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre_completo": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "rol": {
                        "type": "string",
                        "example": "Mesero"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Menu": {
                "required": [
                    "nombre",
                    "precio",
                    "categoria"
                ],
                "properties": {
                    "id_menu": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Hamburguesa Clásica"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 120
                    },
                    "categoria": {
                        "type": "string",
                        "example": "comidas"
                    },
                    "disponible": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "PedidoDetalleInput": {
                "required": [
                    "id_menu",
                    "cantidad",
                    "precio_unitario"
                ],
                "properties": {
                    "id_menu": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "cantidad": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 2
                    },
                    "precio_unitario": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 120
                    }
                },
                "type": "object"
            },
            "PedidoCreate": {
                "required": [
                    "id_cliente",
                    "id_empleado",
                    "detalles"
                ],
                "properties": {
                    "id_cliente": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "id_empleado": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "numero_mesa": {
                        "type": "string",
                        "example": "A1"
                    },
                    "id_estado": {
                        "description": "Opcional, por defecto 'Recibido'",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PedidoDetalleInput"
                        }
                    }
                },
                "type": "object"
            },
            "PedidoDetalle": {
                "properties": {
                    "id_detalle": {
                        "type": "integer",
                        "format": "int64",
                        "example": 10
                    },
                    "id_pedido": {
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "id_menu": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 2
                    },
                    "precio_unitario": {
                        "type": "number",
                        "format": "float",
                        "example": 120
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 240
                    },
                    "menu": {
                        "$ref": "#/components/schemas/Menu"
                    }
                },
                "type": "object"
            },
            "Pedido": {
                "properties": {
                    "id_pedido": {
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "id_cliente": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "id_empleado": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "id_estado": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "numero_mesa": {
                        "type": "string",
                        "example": "A1"
                    },
                    "fecha_creacion": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-15T12:34:56Z"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 240
                    },
                    "cliente": {
                        "$ref": "#/components/schemas/Cliente"
                    },
                    "empleado": {
                        "$ref": "#/components/schemas/Empleado"
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PedidoDetalle"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}